requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
packaging==23.2
pefile==2023.2.7
pydantic==2.7.0
pydantic_core==2.18.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pywin32-ctypes==0.2.2
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
paho-mqtt
pyserial